+++ apps/parsley/package.json
@@ -32,7 +32,7 @@
     "verify-backend": "./scripts/verify-backend.sh",
     "storybook": "storybook dev -p 6006",
     "storybook:build": "env-cmd -e local -r .env-cmdrc.local.json storybook build",
-    "test": "yarn --cwd='../..' test --selectProjects parsley",
+    "test": "yarn --cwd='../..' test --selectProjects parsley --watchAll=false",
     "postversion": "scripts/push-version.sh"
   },
   "dependencies": {
