+++ apps/parsley/src/components/DetailsMenu/DetailsMenuCard/Toggles/JumpToFailingLineToggle/index.tsx
@@ -15,8 +15,7 @@ const JumpToFailingLineToggle: React.FC<JumpToFailingLineToggleProps> = ({
 }) => {
   const { sendEvent } = usePreferencesAnalytics();
   const { logMetadata } = useLogContext();
-  const { logType } = logMetadata || {};
-  const isTaskLog = logType === LogTypes.EVERGREEN_TASK_LOGS;
+  const isTaskLog = logMetadata?.logType !== LogTypes.EVERGREEN_TASK_LOGS;
 
   return (
     <BaseToggle
